// web/src/components/git-panel/types/types.ts
export type FileStatusCode = 'M' | 'A' | 'D' | 'U';
export type GitStatusFileGroup = 'modified' | 'added' | 'deleted' | 'untracked';
export type ConfirmActionType = 'discard' | 'delete' | 'commit' | 'pull' | 'push' | 'publish' | 'revertLocalCommit' | 'deleteBranch';
// ...
export type GitStatusGroupEntry = {
  key: GitStatusFileGroup;
  status: FileStatusCode;
};